Add files via upload
This software was written with the full intent of making sure people who need it got it. It was in iteration of my other studies that EPARA and SIFT was developed due to a physics problem i was attempting to resolve (yeah it worked) this is the fruit of those INITIAL findings. MoveMind hits ~99% accuracy for movement intent and is awesome for BCI apps.